CSP Handling for LiveReload: Completely Remove CSP instead of Updating it.
