+++ lib/runHook.js
@@ -3,7 +3,7 @@ var promiseUtils = require('./promise-util');
 var events = require('./events');
 var path = require('path');
 var fs = require('fs');
-var CSPUpdater = require('./CSPUpdater');
+var CSPRemover = require('./CSPRemover');
 
 // ToDO: Test: run 'cordova run --livereload --device' multiple times in a series
 // ToDO: Test: run 'cordova run --device' after '--livereload' and vice-versa
@@ -77,8 +77,8 @@ module.exports = function(context) {
                     return Q();
                 }).then(function() {
 		    // Update CSP to allow calls to BrowserSync HTTP server
-		    var cspUpdater = new CSPUpdater(path.join(projectRoot, 'www', 'index.html')); 
-		    return cspUpdater.Update();		    
+		    var cspRemover = new CSPRemover(path.join(projectRoot, 'www', 'index.html')); 
+		    return cspRemover.Remove();		    
 		});
             }, function(err) {
                 events.emit('warn', err);
